Revert "Merge pull request #11 from Neha220803/feature-accomodation"
This reverts commit 2ba7daef24891f7a020b8d929d9117532048a6ee, reversing
changes made to ce1a655594d582fa6e66d2dbdba88dbc061d4660.

+++ backend/src/main/java/com/travelease/backend/accommodation/dto/HotelReviewResponse.java
@@ -1,5 +1,6 @@
 package com.travelease.backend.accommodation.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record HotelReviewResponse(
@@ -7,7 +8,7 @@ public record HotelReviewResponse(
         UUID hotelId,
         UUID userId,
         String userName,
-        Integer rating,
+        BigDecimal rating,
         String comment
 ) {
 }
